added tags to test suite
