+++ package.json
@@ -8,7 +8,7 @@
     "rbac:setup": "./build/rbac-setup.sh",
     "check": "prettier --check .",
     "check:fix": "prettier --write .",
-    "test": "npm run test:clean-reports && npm run test:headless && npm run test:merge-reports",
+    "test": "npm run test:clean-reports && npm run test:headless",
     "test:api": "jest --colors",
     "test:clean-reports": "rm -rf ./results",
     "test:debug": "cross-env NODE_ENV=debug ./start-tests.sh",
@@ -30,9 +30,9 @@
     "npm": "6.9.0"
   },
   "dependencies": {
-    "axios": "^0.21.4",
+    "axios": "^0.22.0",
     "del": "^6.0.0",
-    "jest": "^27.2.3",
+    "jest": "^27.2.4",
     "jest-junit": "^12.3.0",
     "log4js": "^6.3.0",
     "nconf": "^0.11.3",
@@ -55,7 +55,7 @@
     "junit-report-merger": "^3.0.2",
     "mocha": "^9.1.2",
     "mocha-junit-reporter": "^2.0.0",
-    "mochawesome": "^6.2.2",
+    "mochawesome": "^6.3.0",
     "mochawesome-merge": "^4.2.0",
     "concurrently": "^6.2.2",
     "prettier": "^2.4.1",
