+++ tests/cypress/scripts/cliHelper.js
@@ -83,9 +83,9 @@ export const cliHelper = {
   setup: (modes) => {
     modes.forEach((mode) => {
       if (!mode.skip) {
-        describe(`Search: Create resource in ${mode.label} Cluster`, function () {
+        describe(`Search: Create resource in ${mode.label} Cluster`, { tags: ['@canary'] }, function () {
           // Log into the hub and managed cluster with the oc command to create the resources.
-          context(`prereq: create resource with oc command`, function () {
+          context(`prereq: create resource with oc command`, { tags: ['@required'] }, function () {
             it(`[P1][Sev1][${squad}] should log into ${mode.label.toLocaleLowerCase()} cluster`, function () {
               cliHelper.login(mode.label)
             })
