+++ tests/cypress/tests/03-create/01-create-resource.spec.js
@@ -23,29 +23,26 @@ clusterModes.forEach((clusterMode) => {
     return
   }
 
-  describe(
-    'RHACM4K-1233: Search: Search in ' + clusterMode.label + ' Cluster',
-    function () {
-      context('prereq: user should log into the ACM console', function () {
-        it(`[P1][Sev1][${squad}] should login`, function () {
-          cy.login()
-        })
+  describe('RHACM4K-1233: Search: Search in ' + clusterMode.label + ' Cluster', { tags: [] }, function () {
+    context('prereq: user should log into the ACM console', { tags: ['@required'] }, function () {
+      it(`[P1][Sev1][${squad}] should login`, function () {
+        cy.login()
+      })
+    })
+
+    // For now, we are only clicking the button within the hub cluster header.
+    context('search resources: verify create resource in search', { tags: [] }, function () {
+      beforeEach(function () {
+        searchPage.whenGoToSearchPage()
       })
 
-      // For now, we are only clicking the button within the hub cluster header.
-      context('search resources: verify create resource in search', () => {
-        beforeEach(function () {
-          searchPage.whenGoToSearchPage()
-        })
-
-        it(`[P2][Sev2][${squad}] should load the search page`, function () {
-          searchPage.shouldLoad()
-        })
-
-        it(`[P2][Sev2][${squad}] should have link to resource creation page`, function () {
-          searchPage.whenCreateResourceObject()
-        })
+      it(`[P2][Sev2][${squad}] should load the search page`, function () {
+        searchPage.shouldLoad()
       })
-    }
-  )
+
+      it(`[P2][Sev2][${squad}] should have link to resource creation page`, function () {
+        searchPage.whenCreateResourceObject()
+      })
+    })
+  })
 })
