+++ tests/cypress/tests/05-delete/01-delete-resources.spec.js
@@ -32,24 +32,21 @@ clusterModes.forEach((clusterMode) => {
     return
   }
 
-  describe(
-    'RHACKM4K-726: Search: Search in ' + clusterMode.label + ' Cluster',
-    function () {
+  describe('RHACKM4K-726: Search: Search in ' + clusterMode.label + ' Cluster', { tags: ['@canary', '@rosa'] }, function () {
       before(function () {
         clusterMode.valueFn().as('clusterName')
       })
 
       // Logging into the hub cluster UI.
       if (clusterMode.label !== 'Managed') {
-        context('prereq: user should log into the ACM console', function () {
+        context('prereq: user should log into the ACM console', { tags: ['@required'] }, function () {
           it(`[P1][Sev1][${squad}] should login`, function () {
             cy.login()
           })
         })
       }
 
-      context(
-        'search resource: verify delete function in search result',
+      context('search resource: verify delete function in search result', { tags: ['@bvt'] },
         function () {
           beforeEach(function () {
             searchPage.whenGoToSearchPage()
