+++ tests/cypress/tests/12-rbac/01-view-overview-with-rbacuser.spec.js
@@ -17,27 +17,27 @@ const rbac_users = [
 const password = Cypress.env('OPTIONS_HUB_PASSWORD')
 const IDP = 'search-e2e-htpasswd'
 
-describe(
-  `RBAC users to read the Overview page`,
-  { tags: '@rbac' },
+describe('RBAC users to read the Overview page', { tags: ['@canary', '@rosa'] },
   function () {
     const overviewPagePolarionIDs = ['731', '921', '919', '920']
 
     rbac_users.forEach((user, index) => {
       var roleAccess = user.split('-')
 
-      it(`RHACM4K-${overviewPagePolarionIDs[index]}[P1][Sev1][${squad}] Login: ${user} user`, function () {
-        cy.login(user, password, IDP)
-      })
-
-      it(`RHACM4K-${overviewPagePolarionIDs[index]}[P2][Sev2][${squad}] As an user with name ${user} with ${roleAccess[3]}-role-binding of default ${roleAccess[2]} role, the user can read the Overview page.`, function () {
-        overviewPage.whenGoToOverviewPage()
-        overviewPage.shouldLoad()
-        overviewPage.shouldHaveLinkToSearchPage()
-      })
-
-      it(`RHACM4K-${overviewPagePolarionIDs[index]}[P1][Sev1][${squad}] Logout: ${user} user`, function () {
-        cy.logout()
+      context(`verify: read action for user ${user} with ${roleAccess[2]} role`, { tags: ['@bvt', '@rbac'] }, function () {
+        it(`RHACM4K-${overviewPagePolarionIDs[index]}[P1][Sev1][${squad}] Login: ${user} user`, function () {
+          cy.login(user, password, IDP)
+        })
+  
+        it(`RHACM4K-${overviewPagePolarionIDs[index]}[P2][Sev2][${squad}] As an user with name ${user} with ${roleAccess[3]}-role-binding of default ${roleAccess[2]} role, the user can read the Overview page.`, function () {
+          overviewPage.whenGoToOverviewPage()
+          overviewPage.shouldLoad()
+          overviewPage.shouldHaveLinkToSearchPage()
+        })
+  
+        it(`RHACM4K-${overviewPagePolarionIDs[index]}[P1][Sev1][${squad}] Logout: ${user} user`, function () {
+          cy.logout()
+        })
       })
     })
   }
